Changes for WIN compilation
Former-commit-id: b0bdaa60a570110fdee158e9c532b3508f3333a8

+++ src/cpp/reader/WriterProxy.cpp
@@ -269,11 +269,11 @@ bool WriterProxy::add_unknown_changes(SequenceNumber_t& seq)
 {
 
 	boost::lock_guard<WriterProxy> guard(*this);
-	uint64_t n_to_add;
+	uint32_t n_to_add;
 	SequenceNumber_t maxseqNum = max_seq_num();
 	if((maxseqNum+1) < (seq)) //if the maximum plus one is less than our seqNum we need to add
 	{
-		n_to_add = (seq.to64long() - maxseqNum.to64long() -1);
+		n_to_add = (uint32_t)(seq.to64long() - maxseqNum.to64long() -1);
 		pDebugInfo("WriterProxy:add_unknown_changes: up to: "<<seq.to64long() << " || adding " << n_to_add << " changes."<<endl;);
 		while(n_to_add>0)
 		{
